launchpad #1321919: [odoo] cured masking of python std lib by addons
Odoo does not need the addons to be on the path : it actually adds them. 

On the other hand,having the addons on the path from the
very beginning creates a masking of the 'calendar' module by Odoo addon of the
same name, which prevents startup.

It is probably not possible to directly import 'calendar' any more, since it's
been loaded in sys.modules during the startup sequence and because openerp
module loader does lookup in sys.modules first (for legacy direct import
without proper 'openerp.addons' namespacing), but that kind of problem is not
specific to the recipe, meaning that the Odoo team will fix it if it's
actually harmful to them.