Providing environments initialization in runtime.session
This needs further testing, in particular with older OpenERP/Odoo versions
Besides, the current Odoo master version that this has been developped
against calls itself 9.0.0 alpha, one needs to check if the problem
exists for the 8.0 branch.

At this point, I'm not really sure of what happens for a script that would
spawn several threads.