launchpad #1343518: previous commit precautions where tautological
Indeed, it added 'addons' to sys.path only if it does not exist.
And we know that at least on recent 7.0, it's a big problem that it actually
contains addons.

In other words, I was over cautious, we have to bite the bullet.
Keeping the ``is_git_layout`` boolean, though, it may prove useful for other
reasons.