Fix sporatic bzr test failure
Apparently, now launchpad resolves lp: addresses as bzr+ssh if the
project does not exist (probably to ease creation), but this is refused if
launchpad-login is not set, and that depends on the buildslave running the
test. This change assumes that lp:anybox.recipe.openerp won't disappear, which
is a safe bet for the medium term.