[buildbot] reorg, cfgs for a.r.openerp in related branch
Cherry picked changes done for now in master. The plan is to have for now
a single MANIFEST for all variants, staying in master branch.
But the cfg files for each variant should lie in the corresponding branch only,
to avoid the confusion of having concurrent versions, only one being active.