Ensuring that pip is provided to integrations tests
First merge of the branch for Github #43 showed that these tests actually
relied on a side effect of zc.buildout machinery.
Indeed, buildout automatically adds the location of the setuptools
distribution, and in case that's a whole (virtualenv) site-packages, it also
provides pip. But there are lots of situations where setuptools would be from
its own, separate location (egg installed by setup.py develop --upgrade)

We actually do the same for pip, with the drawback that we
actually make exposure of the whole site-packages more frequent in those
integration tests, but that's safer than the alternative, that'd have been
to lay an egg manually built from the available pip.

We are sure that pip is available, because it's now a requirement.
The new feature (apply-requirements-file) does not work for older pip versions
anyway (seen with 1.1)