[Travis] reproduced and analyzed install problem (more below)
if 'wheel' is available, then pip uses bdist_wheel, then installs
the unpacked produced wheel. But with zc.recipe.egg, it fails where
it checks the wheel version (ironically, since it just built it).
It seems to be related to the name itself, which produces a confusion with
an egg. Hints to that below:

pip code wheel.wheel_version():

  wheel_data = dist.get_metadata('WHEEL')

pdb session:

  (Pdb) dist.get_metadata('WHEEL')
  *** IOError: [Errno 2] No such file or directory: '/tmp/pip-build-bwjdgA/zc.recipe.egg/EGG-INFO/WHEEL'
  (Pdb) dist
  zc.recipe [unknown version] (/tmp/pip-build-bwjdgA/zc.recipe.egg)

whereas of course:

 ls /tmp/pip-build-bwjdgA/zc.recipe.egg
 zc  zc.recipe.egg-2.0.3-py2.7-nspkg.pth  zc.recipe.egg-2.0.3.dist-info

Problem occurred with all pip versions tested at the time of this commit and
analysed under pip 8.1.1 (setuptools 12.0.5)