Restoring a truly fake babel for cases it's useful
and renaming BabelFake as FictiveDist

The recipe still has hardcoded dependencies to Babel,
(mostly for historical reasons, but we don't want to clean
that up right now), so a fake babel posing really as Babel is
needed.

Had not been spotted in the course of the previous refactor,
because the change of name induced a change in cleanups, and
I had a Babel.egg-info shadowing the problem.